refactor(agents): split prompt pending helpers
